Update the pipeline to allow different kinds of tests to run
We are going to have a caterogy of tests that run against the live service that
we don't want to run in our CI as they require secrets in order to connect to
the service.

However, those tests should run on the developers machine as they can provide
those secrets.

This change is ported over from the lab repo.